bugfix: fix HttpAutoConfiguration always instantiation in springboot env (#3097)
