--- integration/http/src/main/resources/META-INF/spring.factories
@@ -1,3 +1,0 @@
-# Auto Configuration
-org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-io.seata.integration.http.HttpAutoConfiguration
