+++ seata-spring-boot-starter/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,4 @@
 # Auto Configure
-org.springframework.boot.autoconfigure.EnableAutoConfiguration=io.seata.spring.boot.autoconfigure.SeataAutoConfiguration
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
+io.seata.spring.boot.autoconfigure.SeataAutoConfiguration,\
+io.seata.spring.boot.autoconfigure.HttpAutoConfiguration
